+++ src/core/multiconductor.jl
@@ -161,9 +161,19 @@ Base.rad2deg(a::MultiConductorMatrix) = MultiConductorMatrix(map(rad2deg, a.valu
 Base.deg2rad(a::MultiConductorVector) = MultiConductorVector(map(deg2rad, a.values))
 Base.deg2rad(a::MultiConductorMatrix) = MultiConductorMatrix(map(deg2rad, a.values))
 
-JSON.lower(mcv::PowerModels.MultiConductorValue) = Dict("values"=>mcv.values, "type"=>string(typeof(mcv)))
+JSON.lower(mcv::PowerModels.MultiConductorValue) = Dict("values"=>[eltype(mcv) != String && (isinf(v) || isnan(v)) ? string(v) : v for v in mcv.values], "type"=>string(typeof(mcv)))
 function JSON.show_json(io::JSON.StructuralContext, s::JSON.CommonSerialization, p::PowerModels.MultiConductorValue)
-    a = Dict("values"=>p.values, "type"=>string(typeof(p)))
+    if eltype(p) != String
+        if isa(p, MultiConductorMatrix)
+            values = [[!isa(v, String) && isinf(v) || isnan(v) ? string(v) : v for v in row] for row in p.values]
+        elseif isa(p, MultiConductorVector)
+            values = [!isa(v, String) && isinf(v) || isnan(v) ? string(v) : v for v in p.values]
+        end
+    else
+        values = p.values
+    end
+
+    a = Dict("values"=>values, "type"=>string(typeof(p)))
     JSON.begin_object(io)
     for kv in a
         JSON.show_pair(io, s, kv)