+++ src/io/json.jl
@@ -14,7 +14,8 @@ function _parse_mcv!(pm_data)
                    "Int"=>Int64,
                    "Real"=>Real,
                    "Complex"=>Complex,
-                   "String"=>String)
+                   "String"=>String,
+                   "Bool"=>Bool)
 
     for (comp_type, comp_items) in pm_data
         if isa(comp_items, Dict)
@@ -23,9 +24,19 @@ function _parse_mcv!(pm_data)
                     if isa(value, Dict) && haskey(value, "values") && haskey(value, "type")
                         element_type = match(r"MultiConductor(?:Vector|Matrix){([a-zA-Z]+)\d*}", value["type"]).captures[1]
                         if startswith(value["type"], "MultiConductorVector") || startswith(value["type"], "PowerModels.MultiConductorVector")
-                            pm_data[comp_type][n][field] = PowerModels.MultiConductorVector(convert(Array{eltypes[element_type]}, value["values"]))
+                            if element_type != "String"
+                                values = [isa(v, AbstractString) ? parse(eltypes[element_type], v) : v for v in value["values"]]
+                            else
+                                values = value["values"]
+                            end
+                            pm_data[comp_type][n][field] = PowerModels.MultiConductorVector(convert(Array{eltypes[element_type]}, values))
                         elseif startswith(value["type"], "MultiConductorMatrix") || startswith(value["type"], "PowerModels.MultiConductorMatrix")
-                            pm_data[comp_type][n][field] = PowerModels.MultiConductorMatrix(convert(Array{eltypes[element_type]}, hcat(value["values"]...)))
+                            if element_type != "String"
+                                values = [[isa(v, AbstractString) ? parse(eltypes[element_type], v) : v for v in row] for row in value["values"]]
+                            else
+                                values = value["values"]
+                            end
+                            pm_data[comp_type][n][field] = PowerModels.MultiConductorMatrix(convert(Array{eltypes[element_type]}, hcat(values...)))
                         end
                     end
                 end